Dirty hack to start playing song.
iTunes has a bug in that searching for a song using the applescript API will
highlight that search entry in iTunes but not select it. If it were selected,
all we'd have to do is send a button press. Unfortunately, because the content
is not selected we have to work around and read the elements of the UI. This
requires extra accessibility permissions.